requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
click @ file:///work/perseverance-python-buildout/croot/click_1728387069202/work
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe @ file:///croot/markupsafe_1736366905094/work
mysql-connector-python==9.1.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.44.0
